// packages/beecli/sdk/beetrade-v2/models/CreateBotRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateBotRequest
 */
export interface CreateBotRequest {
    /**
     * 
     * @type {number}
     * @memberof CreateBotRequest
     */
    botTypeId: number;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof CreateBotRequest
     */
    configData: { [key: string]: any; };
}

/**
 * Check if a given object implements the CreateBotRequest interface.
 */
export function instanceOfCreateBotRequest(value: object): value is CreateBotRequest {
    if (!('botTypeId' in value) || value['botTypeId'] === undefined) return false;
    if (!('configData' in value) || value['configData'] === undefined) return false;
    return true;
}

export function CreateBotRequestFromJSON(json: any): CreateBotRequest {
    return CreateBotRequestFromJSONTyped(json, false);
}

export function CreateBotRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateBotRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'botTypeId': json['bot_type_id'],
        'configData': json['config_data'],
    };
}

export function CreateBotRequestToJSON(json: any): CreateBotRequest {
    return CreateBotRequestToJSONTyped(json, false);
}

export function CreateBotRequestToJSONTyped(value?: CreateBotRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'bot_type_id': value['botTypeId'],
        'config_data': value['configData'],
    };
}