// packages/beecli/sdk/beetrade-v2/models/UpdateBotRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BotStatus } from './BotStatus';
import {
    BotStatusFromJSON,
    BotStatusFromJSONTyped,
    BotStatusToJSON,
    BotStatusToJSONTyped,
} from './BotStatus';

/**
 * 
 * @export
 * @interface UpdateBotRequest
 */
export interface UpdateBotRequest {
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof UpdateBotRequest
     */
    configData?: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof UpdateBotRequest
     */
    name?: string;
    /**
     * 
     * @type {BotStatus}
     * @memberof UpdateBotRequest
     */
    status?: BotStatus;
}



/**
 * Check if a given object implements the UpdateBotRequest interface.
 */
export function instanceOfUpdateBotRequest(value: object): value is UpdateBotRequest {
    return true;
}

export function UpdateBotRequestFromJSON(json: any): UpdateBotRequest {
    return UpdateBotRequestFromJSONTyped(json, false);
}

export function UpdateBotRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateBotRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'configData': json['config_data'] == null ? undefined : json['config_data'],
        'name': json['name'] == null ? undefined : json['name'],
        'status': json['status'] == null ? undefined : BotStatusFromJSON(json['status']),
    };
}

export function UpdateBotRequestToJSON(json: any): UpdateBotRequest {
    return UpdateBotRequestToJSONTyped(json, false);
}

export function UpdateBotRequestToJSONTyped(value?: UpdateBotRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'config_data': value['configData'],
        'name': value['name'],
        'status': BotStatusToJSON(value['status']),
    };
}