// packages/beecli/sdk/beetrade-v2/models/BotBacktestRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BotBacktestRequest
 */
export interface BotBacktestRequest {
    /**
     * 
     * @type {number}
     * @memberof BotBacktestRequest
     */
    brokerageId: number;
    /**
     * 
     * @type {number}
     * @memberof BotBacktestRequest
     */
    cash: number;
    /**
     * 
     * @type {string}
     * @memberof BotBacktestRequest
     */
    endDate: string;
    /**
     * 
     * @type {string}
     * @memberof BotBacktestRequest
     */
    startDate: string;
}

/**
 * Check if a given object implements the BotBacktestRequest interface.
 */
export function instanceOfBotBacktestRequest(value: object): value is BotBacktestRequest {
    if (!('brokerageId' in value) || value['brokerageId'] === undefined) return false;
    if (!('cash' in value) || value['cash'] === undefined) return false;
    if (!('endDate' in value) || value['endDate'] === undefined) return false;
    if (!('startDate' in value) || value['startDate'] === undefined) return false;
    return true;
}

export function BotBacktestRequestFromJSON(json: any): BotBacktestRequest {
    return BotBacktestRequestFromJSONTyped(json, false);
}

export function BotBacktestRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): BotBacktestRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageId': json['brokerageId'],
        'cash': json['cash'],
        'endDate': json['endDate'],
        'startDate': json['startDate'],
    };
}

export function BotBacktestRequestToJSON(json: any): BotBacktestRequest {
    return BotBacktestRequestToJSONTyped(json, false);
}

export function BotBacktestRequestToJSONTyped(value?: BotBacktestRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageId': value['brokerageId'],
        'cash': value['cash'],
        'endDate': value['endDate'],
        'startDate': value['startDate'],
    };
}