// packages/beecli/sdk/beetrade-v2/models/BotPaperTradingRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BotPaperTradingRequest
 */
export interface BotPaperTradingRequest {
    /**
     * 
     * @type {number}
     * @memberof BotPaperTradingRequest
     */
    brokerageId: number;
    /**
     * 
     * @type {number}
     * @memberof BotPaperTradingRequest
     */
    cash: number;
}

/**
 * Check if a given object implements the BotPaperTradingRequest interface.
 */
export function instanceOfBotPaperTradingRequest(value: object): value is BotPaperTradingRequest {
    if (!('brokerageId' in value) || value['brokerageId'] === undefined) return false;
    if (!('cash' in value) || value['cash'] === undefined) return false;
    return true;
}

export function BotPaperTradingRequestFromJSON(json: any): BotPaperTradingRequest {
    return BotPaperTradingRequestFromJSONTyped(json, false);
}

export function BotPaperTradingRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): BotPaperTradingRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageId': json['brokerageId'],
        'cash': json['cash'],
    };
}

export function BotPaperTradingRequestToJSON(json: any): BotPaperTradingRequest {
    return BotPaperTradingRequestToJSONTyped(json, false);
}

export function BotPaperTradingRequestToJSONTyped(value?: BotPaperTradingRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageId': value['brokerageId'],
        'cash': value['cash'],
    };
}