// packages/beecli/sdk/beetrade-v2/models/BotLiveTradingRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BotLiveTradingRequest
 */
export interface BotLiveTradingRequest {
    /**
     * 
     * @type {number}
     * @memberof BotLiveTradingRequest
     */
    brokerageAccountId: number;
}

/**
 * Check if a given object implements the BotLiveTradingRequest interface.
 */
export function instanceOfBotLiveTradingRequest(value: object): value is BotLiveTradingRequest {
    if (!('brokerageAccountId' in value) || value['brokerageAccountId'] === undefined) return false;
    return true;
}

export function BotLiveTradingRequestFromJSON(json: any): BotLiveTradingRequest {
    return BotLiveTradingRequestFromJSONTyped(json, false);
}

export function BotLiveTradingRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): BotLiveTradingRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageAccountId': json['brokerageAccountId'],
    };
}

export function BotLiveTradingRequestToJSON(json: any): BotLiveTradingRequest {
    return BotLiveTradingRequestToJSONTyped(json, false);
}

export function BotLiveTradingRequestToJSONTyped(value?: BotLiveTradingRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageAccountId': value['brokerageAccountId'],
    };
}